Adding source_id corrections and unit test.... class coverage 97%, method coverage 90%, line coverage 84%

+++ src/main/java/ca/homedepot/preference/model/FileInboundStgTable.java
@@ -37,7 +37,7 @@ public class FileInboundStgTable
 	String email_pref_garden_club;
 	String email_pref_pro;
 	String email_pref_new_mover;
-	Long source_id;
+	BigDecimal source_id;
 	String cell_sms_flag;
 	String customer_nbr;
